Use UserName in all places with user names
