+++ src/components/more-comments-wrapper.jsx
@@ -1,8 +1,10 @@
 import React from 'react'
 
+import {preventDefault} from '../utils'
+
 export default (props) => (
   <div className="more-comments-wrapper">
-    <a className="more-comments" onClick={(e) => {e.preventDefault(); props.showMoreComments()}}>
+    <a className="more-comments" onClick={preventDefault(()=>props.showMoreComments())}>
       {`${props.omittedComments}`} more comments
     </a>
   </div>
